+++ tests/nat/integration/test_nat_adaptors.py
@@ -19,12 +19,10 @@ from llama_index.core.llms import ChatResponse
 from nat.builder.framework_enum import LLMFrameworkEnum
 from nat.builder.workflow_builder import WorkflowBuilder
 
-from datarobot_genai.nat_adaptors.datarobot_llm_clients import (
-    datarobot_llm_gateway_langchain,  # noqa: F401
-)
-from datarobot_genai.nat_adaptors.datarobot_llm_providers import DataRobotLLMDeploymentModelConfig
-from datarobot_genai.nat_adaptors.datarobot_llm_providers import DataRobotLLMGatewayModelConfig
-from datarobot_genai.nat_adaptors.datarobot_llm_providers import DataRobotNIMModelConfig
+from datarobot_genai.nat.datarobot_llm_clients import datarobot_llm_gateway_langchain  # noqa: F401
+from datarobot_genai.nat.datarobot_llm_providers import DataRobotLLMDeploymentModelConfig
+from datarobot_genai.nat.datarobot_llm_providers import DataRobotLLMGatewayModelConfig
+from datarobot_genai.nat.datarobot_llm_providers import DataRobotNIMModelConfig
 
 
 async def test_datarobot_llm_gateway_langchain():
